+++ src/jbookreader/rendering/IFont.java
@@ -1,7 +1,7 @@
 package jbookreader.rendering;
 
 public interface IFont {
-	int getSpaceWidth();
+	float getSpaceWidth();
 
 	String getFamily();
 	int getSize();
